move expected_arch to testing
