+++ tests/python/test_test.py
@@ -14,7 +14,7 @@ import taichi as ti
 
 @ti.test()
 def test_all_archs():
-    assert ti.cfg.arch in ti.expected_archs()
+    assert ti.cfg.arch in ti.testing.expected_archs()
 
 
 @ti.test(arch=ti.cpu)
